community_usecase/requirements-analysis-assistant/requirements.txt
# Core dependencies
loguru>=0.7.0
pydantic>=2.11.0
aiohttp>=3.8.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
jinja2>=3.0.0
pyyaml>=6.0.0
markdown>=3.4.0
httpx>=0.27.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
pre-commit>=3.0.0

# Additional dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
prompt_toolkit>=3.0.0
websockets>=10.0
python-multipart>=0.0.5
aiofiles>=0.8.0
pydantic-settings>=2.1.0
tiktoken>=0.6.0
loguru>=0.7.2
gradio>=4.19.2

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# CLI
click>=8.0.0

# Utils
python-dotenv>=1.0.1
requests>=2.26.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.14
mkdocstrings>=0.24.1
mkdocstrings-python>=1.9.0 